Enum moved to enums package
